Extract NumberConstraintTest
